+++ app/src/main/java/com/hedvig/app/feature/marketpicker/Market.kt
@@ -1,6 +1,12 @@
 package com.hedvig.app.feature.marketpicker
 
+import android.content.Context
+import androidx.fragment.app.FragmentManager
 import com.hedvig.app.R
+import com.hedvig.app.authenticate.AuthenticateDialog
+import com.hedvig.app.feature.chat.ui.ChatActivity
+import com.hedvig.app.feature.webonboarding.WebOnboardingActivity
+import com.hedvig.app.feature.zignsec.ZignSecAuthenticationActivity
 
 enum class Market {
     SE,
@@ -20,6 +26,25 @@ enum class Market {
             NO -> R.string.norway
             DK -> R.string.denmark
         }
+
+    fun openAuth(context: Context, fragmentManager: FragmentManager) {
+        when (this) {
+            SE -> {
+                AuthenticateDialog().show(fragmentManager, AuthenticateDialog.TAG)
+            }
+            NO, DK -> {
+                context.startActivity(ZignSecAuthenticationActivity.newInstance(context))
+            }
+        }
+    }
+
+    fun onboarding(context: Context) = when (this) {
+        SE -> ChatActivity.newInstance(context)
+            .apply { putExtra(ChatActivity.EXTRA_SHOW_RESTART, true) }
+        NO, DK -> {
+            WebOnboardingActivity.newInstance(context)
+        }
+    }
 
     companion object {
         const val MARKET_SHARED_PREF = "MARKET_SHARED_PREF"
