+++ app/src/main/java/com/hedvig/app/feature/zignsec/ZignSecAuthRepository.kt
@@ -1,15 +1,21 @@
-package com.hedvig.app.feature.norway
+package com.hedvig.app.feature.zignsec
 
 import com.apollographql.apollo.coroutines.toDeferred
 import com.apollographql.apollo.coroutines.toFlow
 import com.hedvig.android.owldroid.graphql.AuthStatusSubscription
+import com.hedvig.android.owldroid.graphql.DanishAuthMutation
 import com.hedvig.android.owldroid.graphql.NorwegianBankIdAuthMutation
 import com.hedvig.app.ApolloClientWrapper
 
-class NorwegianAuthenticationRepository(
+class ZignSecAuthRepository(
     private val apolloClientWrapper: ApolloClientWrapper
 ) {
-    fun startAuthAsync() = apolloClientWrapper
+    fun startDanishAuthAsync() = apolloClientWrapper
+        .apolloClient
+        .mutate(DanishAuthMutation())
+        .toDeferred()
+
+    fun startNorwegianAuthAsync() = apolloClientWrapper
         .apolloClient
         .mutate(NorwegianBankIdAuthMutation())
         .toDeferred()
