+++ app/src/main/res/layout/activity_zign_sec_authentication.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".feature.denmark.ZignSecAuthenticationActivity">
+    tools:context=".feature.zignsec.ZignSecAuthenticationActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
